move dataSource and struct Post into new file Post commit
